index property changed to uint8

+++ golang-backend/internal/domain/project_priority.go
@@ -4,7 +4,7 @@ type ProjectPriority struct {
 	ID       string
 	Name     string
 	HexColor string
-	Index    uint
+	Index    uint8
 }
 
 type SearchProjectPrioritiesFilter struct {
