+++ golang-backend/internal/domain/task_priority.go
@@ -4,7 +4,7 @@ type TaskPriority struct {
 	ID       string
 	Name     string
 	HexColor string
-	Index    uint
+	Index    uint8
 }
 
 type SearchTaskPrioritiesFilter struct {
