+++ golang-backend/internal/repositories/projectstatusrepository/dto.go
@@ -4,7 +4,7 @@ type projectStatusDTO struct {
 	ID           string `db:"id"`
 	Name         string `db:"name"`
 	HexColor     string `db:"item_hex_color"`
-	Index        uint   `db:"item_index"`
+	Index        uint8  `db:"item_index"`
 	FlagsBitmask uint64 `db:"flags_bitmask"`
 }
